Extract subpoject for shared kernel
