+++ Shared Kernel/EventBus/BlockEventConsumer.swift
