+++ Shared Kernel/EventBus/EventBus+Convenience.swift
