+++ Shared Kernel/EventBus/EventConsumer.swift
