+++ Shared Kernel/EventBus/EventHandlerRegistration.swift
