+++ Shared Kernel/EventBus/Info.plist
