+++ Shared Kernel/TestUtilities/Random Data Generation/Foundation/Character+RandomValueProviding.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension Character: RandomValueProviding {
+
+    public static var random: Character {
+        guard let character = String.random.first else {
+            fatalError("Failed to produce random Character")
+        }
+        
+        return character
+    }
+
+}
