+++ Shared Kernel/TestUtilities/Random Data Generation/Foundation/Data+RandomValueProviding.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension Data: RandomValueProviding {
+
+    public static var random: Data {
+        guard let data = String.random.data(using: .utf8) else {
+            fatalError("Failed to produce random data")
+        }
+        
+        return data
+    }
+
+}
