--- TestUtilities/Random Data Generation/Foundation/Character+RandomValueProviding.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension Character: RandomValueProviding {
-
-    public static var random: Character {
-        return unwrap(String.random.first)
-    }
-
-}
