--- TestUtilities/Random Data Generation/Foundation/Data+RandomValueProviding.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension Data: RandomValueProviding {
-
-    public static var random: Data {
-        return unwrap(String.random.data(using: .utf8))
-    }
-
-}
