--- TestUtilities/Random Data Generation/Foundation/URL+RandomValueProviding.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension URL: RandomValueProviding {
-
-    public static var random: URL {
-        return unwrap(URL(string: "https://\(String.random)"))
-    }
-
-}
